Update SourceID type to match sourceid package

+++ internal/parseline/record.go
@@ -3,7 +3,7 @@ package parseline
 // A Record represents a single row in the database
 type Record struct {
 	Source   string
-	SourceID uint32
+	SourceID int64
 	Username string
 	Email    string
 	EmailRev string
